requirements.txt
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.4.0
django-decouple==2.1
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
phonenumbers==8.12.7
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.3
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
